+++ packages/terra-date-time-picker/tests/wdio/date-time-picker-spec.js
@@ -279,9 +279,9 @@ Terra.describeViewports('DateTimePicker', ['tiny', 'large'], () => {
 
     Terra.it.matchesScreenshot('0');
 
-    it('Dismissed the modal manager after a click', () => {
+    it('Select a date from the picker', () => {
       browser.click('[class*="button"]');
-      browser.click('[aria-label="day-27"]');
+      browser.click('.react-datepicker__week > *:nth-child(2)');
     });
 
     Terra.it.matchesScreenshot('1');
@@ -305,6 +305,7 @@ Terra.describeViewports('DateTimePicker', ['tiny', 'large'], () => {
   describe('Clears the default date and time on date input focus when default date is excluded', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-default-date-excluded');
+      browser.refresh();
       // Removes the blinking cursor to prevent screenshot mismatches.
       browser.execute('document.querySelector(\'input[name="terra-date-input"]\').style.caretColor = "transparent";');
       browser.click('input[name="terra-date-input"]');
@@ -316,6 +317,7 @@ Terra.describeViewports('DateTimePicker', ['tiny', 'large'], () => {
   describe('Clears the default date and time on hour input focus when default date is excluded', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-default-date-excluded');
+      browser.refresh();
       // Removes the blinking cursor to prevent screenshot mismatches.
       browser.execute('document.querySelector(\'input[name="terra-time-hour-input"]\').style.caretColor = "transparent";');
       browser.click('input[name="terra-time-hour-input"]');
@@ -327,6 +329,7 @@ Terra.describeViewports('DateTimePicker', ['tiny', 'large'], () => {
   describe('Clears the default date and time on minute input focus when default date is excluded', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-default-date-excluded');
+      browser.refresh();
       // Removes the blinking cursor to prevent screenshot mismatches.
       browser.execute('document.querySelector(\'input[name="terra-time-minute-input"]\').style.caretColor = "transparent";');
       browser.click('input[name="terra-time-minute-input"]');
@@ -338,6 +341,7 @@ Terra.describeViewports('DateTimePicker', ['tiny', 'large'], () => {
   describe('Clears the default date and time on date input focus when default date is out of range', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-default-date-out-of-range');
+      browser.refresh();
       // Removes the blinking cursor to prevent screenshot mismatches.
       browser.execute('document.querySelector(\'input[name="terra-date-input"]\').style.caretColor = "transparent";');
       browser.click('input[name="terra-date-input"]');
@@ -349,6 +353,7 @@ Terra.describeViewports('DateTimePicker', ['tiny', 'large'], () => {
   describe('Clears the default date and time on hour input focus when default date is out of range', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-default-date-out-of-range');
+      browser.refresh();
       // Removes the blinking cursor to prevent screenshot mismatches.
       browser.execute('document.querySelector(\'input[name="terra-time-hour-input"]\').style.caretColor = "transparent";');
       browser.click('input[name="terra-time-hour-input"]');
@@ -360,6 +365,7 @@ Terra.describeViewports('DateTimePicker', ['tiny', 'large'], () => {
   describe('Clears the default date and time on minute input focus when default date is out of range', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-date-time-picker/date-time-picker/date-time-picker-default-date-out-of-range');
+      browser.refresh();
       // Removes the blinking cursor to prevent screenshot mismatches.
       browser.execute('document.querySelector(\'input[name="terra-time-minute-input"]\').style.caretColor = "transparent";');
       browser.click('input[name="terra-time-minute-input"]');