+++ packages/terra-embedded-content-consumer/src/terra-dev-site/test/embedded-content-consumer/consumers/BasicConsumer.test.jsx
@@ -1,12 +1,17 @@
 import React from 'react';
 import { Consumer } from 'xfc';
+import classNames from 'classnames/bind';
 import EmbeddedContentConsumer from '../../../../EmbeddedContentConsumer';
+import styles from './Consumer.module.scss';
+
+const cx = classNames.bind(styles);
 
 Consumer.init();
 
 const BasicConsumer = () => (
   <EmbeddedContentConsumer
-    src="#/raw/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/basic-provider"
+    className={cx('iframe')}
+    src="/#/raw/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/basic-provider"
   />
 );