+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+
+4.5.0 - (July 2, 2019)
+------------------
+### Added
+* Allow terra-date-picker to be a controlled component.
+* Added functionality to announce the pre-selected date to screen readers when navigating within the date picker.
+
 ### Changed
 * Removed extra inline styles.
 * Renamed test and example SCSS files to end with `.module.scss` extension