+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -7,9 +7,13 @@ Unreleased
 * Added the ability for the user to enter seconds in the time input
 
 ### Changed
-* Screenshots updated for mobile view.
+* Fixed tests to appropriately reset before beginning next test.
 
+4.5.0 - (July 2, 2019)
+------------------
 ### Changed
+* Screenshots updated for mobile view.
+* Update wdio test to work with changes in terra-date-picker
 * Renamed test and example SCSS files to end with `.module.scss` extension
 
 4.4.0 - (June 19, 2019)