+++ packages/terra-infinite-list/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+3.8.0 - (July 2, 2019)
+------------------
 ### Changed
 * Removed extra inline styles.
 * Renamed test and example SCSS files to end with `.module.scss` extension