+++ packages/terra-navigation-prompt/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+1.3.0 - (July 2, 2019)
+------------------
 ### Changed
 * Renamed test and example SCSS files to end with `.module.scss` extension