+++ packages/terra-slide-panel-manager/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 ### Changed
+* Ensure the disclosure discloses as a panel even if a modal manager is present.
+
+5.3.0 - (July 2, 2019)
+------------------
+### Changed
 * Renamed test and example SCSS files to end with `.module.scss` extension
 
 5.2.0 - (June 19, 2019)